requirement.txt
tkinter
google-auth
google-cloud-storage